Make cmd names more clear
