+++ pkg/ntripmqtt/mount.go
@@ -1,4 +1,4 @@
-package ntrip
+package ntripmqtt
 
 import (
 	"fmt"
